move test codes about GiftedChat Component to GiftedChat.test.tsx
